# cmd/docs/swagger.yaml
definitions:
...
  aqary_admin_internal_domain_requests_user.CreateUserReq:
    properties:
      about:
        type: string
      about_arabic:
        type: string
      botim_number:
        type: string
      brn_expiry:
        type: string
      brn_license_issue_date:
        description: brn file ...................................
        type: string
      brn_no:
        description: '? other detail'
        type: string
      city_id:
        type: integer
      community_id:
        type: integer
      company_id:
        type: integer
      country_id:
        description: |-
          Locations []int64 `form:"locations"`
          location
        type: integer
      department_id:
        type: integer
      email:
        type: string
      experience_since:
        type: string
      facebook:
        type: string
      feature:
        type: integer
      first_name:
        type: string
      gender:
        type: integer
      is_show_user_details:
        type: boolean
      is_verified:
        type: boolean
      last_name:
        type: string
      license_expiry_date:
        type: string
      license_registration_date:
        type: string
      linkedin:
        type: string
      nationalities:
        items:
          type: integer
        type: array
      passport_no:
        type: string
      premium:
        type: integer
      primary_phone_number:
        type: string
      role_id:
        type: integer
      secondary_phone_number:
        type: string
      show_botim_number:
        type: boolean
      show_hide_details:
        description: passport pic
        type: boolean
      show_tawasal_number:
        type: boolean
      show_whatsapp_number:
        type: boolean
      spoken_languages:
        items:
          type: integer
        type: array
      standard:
        type: integer
      state_id:
        type: integer
      sub_community_id:
        type: integer
      tawasal_number:
        type: string
      topdeal:
        type: integer
      twitter:
        type: string
      user_type_id:
        type: integer
      username:
        type: string
      whatsapp_number:
        description: numbers
        type: string
    required:
    - department_id
    - email
    - first_name
    - last_name
    - primary_phone_number
    - role_id
    - user_type_id
    - username
    type: object
...
  aqary_admin_internal_domain_requests_user.GetAllOtherUserByCountryRequest:
    properties:
      country_id:
        type: integer
      page_no:
        type: integer
      page_size:
        type: integer
    required:
    - country_id
    - page_no
    - page_size
    type: object
  aqary_admin_internal_domain_requests_user.GetAllUserRequest:
    properties:
      page_no:
        minimum: 1
        type: integer
      page_size:
        minimum: 1
        type: integer
      search:
        type: string
    required:
    - page_no
    - page_size
    type: object
  aqary_admin_internal_domain_requests_user.GetSingleUserReq:
    properties:
      company_id:
        type: integer
    type: object
...
  aqary_admin_internal_domain_requests_user.Request:
    properties:
      company_id:
        type: integer
      company_type:
        type: integer
      is_branch:
        type: boolean
      permissions_id:
        items:
          type: integer
        type: array
      subsection_id:
        items:
          type: integer
        type: array
    required:
    - company_id
    - company_type
    type: object
...
  aqary_admin_internal_domain_responses_user.AllPendingUserOutput:
    properties:
      city:
        type: string
      community:
        type: string
      company_id:
        type: integer
      company_type:
        type: integer
      country:
        type: string
      department:
        type: string
      email:
        type: string
      id:
        type: integer
      is_branch:
        type: boolean
      phone:
        type: string
      role:
        type: string
      state:
        type: string
      sub_community:
        type: string
      username:
        type: string
      verification_status:
        type: string
    type: object
...
  aqary_admin_internal_domain_responses_user.AllUserOutput:
    properties:
      city:
        type: string
      community:
        type: string
      country:
        type: string
      department:
        type: string
      email:
        type: string
      id:
        type: integer
      location:
        type: string
      phone:
        type: string
      role:
        type: string
      state:
        type: string
      status:
        $ref: '#/definitions/aqary_admin_pkg_utils_fn.CustomFormat'
      sub_community:
        type: string
      username:
        type: string
    type: object
...
  aqary_admin_internal_domain_responses_user.UserOutput:
    properties:
      city:
        $ref: '#/definitions/aqary_admin_pkg_utils_fn.CustomFormat'
      community:
        $ref: '#/definitions/aqary_admin_pkg_utils_fn.CustomFormat'
      country:
        $ref: '#/definitions/aqary_admin_pkg_utils_fn.CustomFormat'
      department: {}
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      permission: {}
      phone:
        type: string
      role: {}
      state:
        $ref: '#/definitions/aqary_admin_pkg_utils_fn.CustomFormat'
      sub_community:
        $ref: '#/definitions/aqary_admin_pkg_utils_fn.CustomFormat'
      user_type_id:
        type: integer
      username:
        type: string
    type: object
...
  aqary_admin_internal_domain_sqlc_sqlc.GetRemainingCompanyQuotaRow:
    properties:
      product:
        type: string
      total_no_of_products:
        type: integer
    type: object
...
  aqary_admin_pkg_utils.ErrResponseSwagger:
    properties:
      error:
        example: any error message
        type: string
    type: object
...
  aqary_admin_pkg_utils_fn.CustomFormat:
    properties:
      id:
        type: integer
      label:
        type: string
    type: object
...
info:
  contact: {}
  description: Testing Swagger APIs.
  title: Aqary  APIs
  version: "1.0"
paths:
...
  /api/user/createCompanyUser:
    post:
      consumes:
      - application/json
      description: Create a new company user with the provided details
      parameters:
      - description: User details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/aqary_admin_internal_domain_requests_user.CreateUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
      security:
      - bearerToken: []
      summary: Create a new company user
      tags:
      - company-users
...
  /api/user/deletePendingUser:
    delete:
      description: DeletePendingUser
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
      security:
      - bearerToken: []
      summary: DeletePendingUser
      tags:
      - pending_user
...
  /api/user/getAllOtherUsers:
    get:
      description: GetAllOtherUser
      parameters:
      - description: GetAllUserRequest
        in: body
        name: UserTypeRequest
        required: true
        schema:
          $ref: '#/definitions/aqary_admin_internal_domain_requests_user.GetAllUserRequest'
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/aqary_admin_internal_domain_responses_user.AllUserOutput'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
      security:
      - bearerToken: []
      summary: GetAllOtherUser
      tags:
      - other_user
  /api/user/getAllOtherUsersByCountry:
    get:
      description: GetAllOtherUserByCountry
      parameters:
      - description: GetAllOtherUserByCountryRequest
        in: body
        name: UserTypeRequest
        required: true
        schema:
          $ref: '#/definitions/aqary_admin_internal_domain_requests_user.GetAllOtherUserByCountryRequest'
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/aqary_admin_internal_domain_responses_user.AllUserOutput'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
      security:
      - bearerToken: []
      summary: GetAllOtherUserByCountry
      tags:
      - other_user
  /api/user/getAllPeddingUsers:
    get:
      description: GetAllPendingUser
      parameters:
      - description: GetAllUserRequest
        in: body
        name: UserTypeRequest
        required: true
        schema:
          $ref: '#/definitions/aqary_admin_internal_domain_requests_user.GetAllUserRequest'
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/aqary_admin_internal_domain_responses_user.AllPendingUserOutput'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
      security:
      - bearerToken: []
      summary: GetAllPendingUser
      tags:
      - pending_user
...
  /api/user/getCurrentSubscriptionQuota:
    get:
      consumes:
      - application/json
      description: GetCurrentSubscriptionQuota
      parameters:
      - description: Permission details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/aqary_admin_internal_domain_requests_user.GetSingleUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/aqary_admin_internal_domain_sqlc_sqlc.GetRemainingCompanyQuotaRow'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
      security:
      - bearerToken: []
      summary: GetCurrentSubscriptionQuota
      tags:
      - company-users
...
  /api/user/getSingleUser/{id}:
    get:
      description: GetSingleUser
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/aqary_admin_internal_domain_responses_user.UserOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
      security:
      - bearerToken: []
      summary: GetSingleUser
      tags:
      - users
...
  /api/user/grantCompanyAdminPermissions:
    post:
      consumes:
      - application/json
      description: Grant specified permissions to a company admin
      parameters:
      - description: Permission details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/aqary_admin_internal_domain_requests_user.Request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aqary_admin_pkg_utils.ErrResponseSwagger'
      security:
      - bearerToken: []
      summary: Grant permissions to a company admin
      tags:
      - company-users
...
securityDefinitions:
  bearerToken:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"